Phase 1: Change salary from Decimal to Float

+++ campus_portal/models.py
@@ -57,7 +57,7 @@ class Student(Person):
 class Employee(Person):
     employee_ID = models.IntegerField(primary_key=True, unique=True)
     department = models.CharField(max_length=10)
-    salary = models.DecimalField(max_digits=10, decimal_places=2)
+    salary = models.FloatField()
 
     def __str__(self):
         return self.name
